Updated automated backup system for branches
Automated backups now push old versions of master to branch master-backup, in case something happens to the master branch and we need to revert.  This happens at midnight every Sunday.  Current versions of master will need to be pulled into the branch manually every once in a while, when they have been verified to be good updates.